+++ flink-connector-clickhouse-2.0.0/src/test/java/org/apache/flink/connector/clickhouse/table/ClickHouseDynamicTableSinkTest.java
@@ -1,7 +1,6 @@
 package org.apache.flink.connector.clickhouse.table;
 
 import com.clickhouse.client.api.metadata.TableSchema;
-import com.clickhouse.config.RetryPolicy;
 import com.clickhouse.data.ClickHouseColumn;
 
 import org.apache.flink.connector.clickhouse.sink.ClickHouseClientConfig;
@@ -28,7 +27,7 @@ class ClickHouseDynamicTableSinkTest {
     /** Built the way the factory builds it, minus the network. */
     private static ClickHouseDynamicTableSink sink() {
         ClickHouseClientConfig config = new ClickHouseClientConfig("http://localhost:1", "u", "", "db", "t",
-                Map.of(), Map.of(), RetryPolicy.forever());
+                Map.of(), Map.of(), false);
         RowDataDataMapper mapper = RowDataDataMapper.of(SchemaResolver.resolve(
                 ResolvedSchema.of(Column.physical("id", DataTypes.BIGINT().notNull())),
                 new TableSchema(ClickHouseColumn.parse("id Int64")),